Weight List Adapter

+++ app/src/main/java/com/example/android/weightlosstracker/models/Weight.kt
@@ -8,6 +8,6 @@ import java.util.*
 data class Weight(
     @PrimaryKey
     var id: Long,
-    var weight: Float,
+    var weight: Double,
     var date: Long
 )
